Substantial rewrite to eliminate a potential race condition.
Previously, the UI took a snapshot of the state of all the baboon
agents and the rope ref before painting. Because agent state isn't
modified in a transaction, however, this can lead to inconsistencies
in the captured world state.

Changed things so that the world state is captured in three separate
refs instead. Baboon agents still wrap their location, but agent
updates are not used in the painting logic.